requirements.txt
langchain_community
langchain-groq
langchainhub
chromadb
langchain
langgraph
langchain-anthropic
tavily-python
pandas
pinecone-client
langchain_core
fastapi
uvicorn
sentence-transformers
bs4
pydantic
gunicorn
anthropic
python-dotenv
user-agents
geoip2